# contracts/agent-service/openapi-python-baseline.json
{
  "components": {
    "schemas": {
      "AgentJobAccepted": {
        "additionalProperties": false,
        "properties": {
          "jobId": {
            "minLength": 1,
            "title": "Jobid",
            "type": "string"
          },
          "protocolVersion": {
            "const": "1.0",
            "default": "1.0",
            "title": "Protocolversion",
            "type": "string"
          },
          "runId": {
            "minLength": 1,
            "title": "Runid",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "taskId": {
            "minLength": 1,
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "runId",
          "taskId",
          "status"
        ],
        "title": "AgentJobAccepted",
        "type": "object"
      },
      "AgentJobCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "novelId": {
            "minLength": 1,
            "title": "Novelid",
            "type": "string"
          },
          "protocolVersion": {
            "const": "1.0",
            "title": "Protocolversion",
            "type": "string"
          },
          "runId": {
            "minLength": 1,
            "title": "Runid",
            "type": "string"
          },
          "taskId": {
            "minLength": 1,
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "protocolVersion",
          "runId",
          "taskId",
          "novelId"
        ],
        "title": "AgentJobCancelRequest",
        "type": "object"
      },
      "AgentJobRequest": {
        "additionalProperties": false,
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "jobId": {
            "minLength": 1,
            "title": "Jobid",
            "type": "string"
          },
          "kind": {
            "enum": [
              "writing",
              "portrait",
              "rag",
              "quality",
              "video"
            ],
            "title": "Kind",
            "type": "string"
          },
          "novelId": {
            "minLength": 1,
            "title": "Novelid",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "priority": {
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "protocolVersion": {
            "const": "1.0",
            "title": "Protocolversion",
            "type": "string"
          },
          "runId": {
            "minLength": 1,
            "title": "Runid",
            "type": "string"
          },
          "taskId": {
            "minLength": 1,
            "title": "Taskid",
            "type": "string"
          },
          "userId": {
            "minLength": 1,
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "protocolVersion",
          "jobId",
          "kind",
          "runId",
          "taskId",
          "novelId",
          "userId",
          "priority",
          "payload"
        ],
        "title": "AgentJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JsonValue": {},
...
      "SeedanceRenderError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "SeedanceRenderError",
        "type": "object"
      },
      "SeedanceRenderOutput": {
        "additionalProperties": false,
        "properties": {
          "durationSeconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationseconds"
          },
          "framesPerSecond": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Framespersecond"
          },
          "generateAudio": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generateaudio"
          },
          "lastFrameUrl": {
            "anyOf": [
              {
                "maxLength": 8192,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastframeurl"
          },
          "mediaKind": {
            "enum": [
              "provider_media",
              "simulated_placeholder"
            ],
            "title": "Mediakind",
            "type": "string"
          },
          "ratio": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "resolution": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "usage": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Usage",
            "type": "object"
          },
          "videoUrl": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Videourl",
            "type": "string"
          }
        },
        "required": [
          "videoUrl",
          "mediaKind"
        ],
        "title": "SeedanceRenderOutput",
        "type": "object"
      },
      "SeedanceRenderQueryRequest": {
        "additionalProperties": false,
        "properties": {
          "executionMode": {
            "enum": [
              "simulated",
              "live"
            ],
            "title": "Executionmode",
            "type": "string"
          },
          "novelId": {
            "minLength": 1,
            "title": "Novelid",
            "type": "string"
          },
          "pollCount": {
            "minimum": 1.0,
            "title": "Pollcount",
            "type": "integer"
          },
          "providerTaskId": {
            "minLength": 1,
            "title": "Providertaskid",
            "type": "string"
          },
          "taskId": {
            "minLength": 1,
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "novelId",
          "providerTaskId",
          "executionMode",
          "pollCount"
        ],
        "title": "SeedanceRenderQueryRequest",
        "type": "object"
      },
      "SeedanceRenderQueryResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeedanceRenderError"
              },
              {
                "type": "null"
              }
            ]
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeedanceRenderOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "providerTaskId": {
            "title": "Providertaskid",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "expired",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "providerTaskId",
          "status"
        ],
        "title": "SeedanceRenderQueryResponse",
        "type": "object"
      },
      "SeedanceRenderSubmitRequest": {
        "additionalProperties": false,
        "description": "Core 从冻结清单投影出的单次 Seedance 创建请求。",
        "properties": {
          "durationSeconds": {
            "maximum": 12.0,
            "minimum": 4.0,
            "title": "Durationseconds",
            "type": "integer"
          },
          "executionMode": {
            "enum": [
              "simulated",
              "live"
            ],
            "title": "Executionmode",
            "type": "string"
          },
          "generateAudio": {
            "title": "Generateaudio",
            "type": "boolean"
          },
          "generationMode": {
            "const": "reference",
            "title": "Generationmode",
            "type": "string"
          },
          "inputHash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Inputhash",
            "type": "string"
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "novelId": {
            "minLength": 1,
            "title": "Novelid",
            "type": "string"
          },
          "promptText": {
            "maxLength": 2500,
            "minLength": 1,
            "title": "Prompttext",
            "type": "string"
          },
          "ratio": {
            "enum": [
              "16:9",
              "4:3",
              "1:1",
              "3:4",
              "9:16",
              "21:9",
              "adaptive"
            ],
            "title": "Ratio",
            "type": "string"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/SeedanceRuntimeReference"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "References",
            "type": "array"
          },
          "resolution": {
            "const": "720p",
            "title": "Resolution",
            "type": "string"
          },
          "taskId": {
            "minLength": 1,
            "title": "Taskid",
            "type": "string"
          },
          "watermark": {
            "title": "Watermark",
            "type": "boolean"
          }
        },
        "required": [
          "taskId",
          "novelId",
          "inputHash",
          "generationMode",
          "executionMode",
          "model",
          "promptText",
          "ratio",
          "durationSeconds",
          "resolution",
          "generateAudio",
          "watermark",
          "references"
        ],
        "title": "SeedanceRenderSubmitRequest",
        "type": "object"
      },
      "SeedanceRenderSubmitResponse": {
        "additionalProperties": false,
        "properties": {
          "providerTaskId": {
            "minLength": 1,
            "title": "Providertaskid",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "providerTaskId"
        ],
        "title": "SeedanceRenderSubmitResponse",
        "type": "object"
      },
      "SeedanceRuntimeReference": {
        "additionalProperties": false,
        "description": "仅存在于受签名内部请求中的短时参考图地址。",
        "properties": {
          "assetId": {
            "minLength": 1,
            "title": "Assetid",
            "type": "string"
          },
          "mimeType": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Mimetype",
            "type": "string"
          },
          "ordinal": {
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Ordinal",
            "type": "integer"
          },
          "url": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          },
          "usageRole": {
            "default": "visual_reference",
            "enum": [
              "visual_reference",
              "initial_state",
              "transition_anchor",
              "end_state"
            ],
            "title": "Usagerole",
            "type": "string"
          }
        },
        "required": [
          "ordinal",
          "assetId",
          "mimeType",
          "url"
        ],
        "title": "SeedanceRuntimeReference",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "InkForge 智能体服务",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/runs": {
      "post": {
        "operationId": "submit_run_internal_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentJobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Run"
      }
    },
    "/internal/v1/runs/{job_id}": {
      "delete": {
        "operationId": "cancel_run_internal_v1_runs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentJobCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run"
      }
    },
    "/internal/v1/video/seedance/tasks": {
      "post": {
        "operationId": "submit_seedance_task_internal_v1_video_seedance_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedanceRenderSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedanceRenderSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Seedance Task"
      }
    },
    "/internal/v1/video/seedance/tasks/{provider_task_id}/query": {
      "post": {
        "operationId": "query_seedance_task_internal_v1_video_seedance_tasks__provider_task_id__query_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_task_id",
            "required": true,
            "schema": {
              "title": "Provider Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedanceRenderQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedanceRenderQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Seedance Task"
      }
    }
  }
}